+++ packages/webapp/src/store/single-view-store.ts
@@ -3,13 +3,18 @@ import { persist } from 'zustand/middleware'
 
 export interface SingleViewStore {
   lastId: string
+  lastIndex: number
   showDetails: boolean
   showNavigation: boolean
 
   setLastId(lastId: string): void
+  setLastIndex(lastIndex: number): void
   setShowDetails(show: boolean): void
   setShowNavigation(show: boolean): void
 }
+
+const excludeStateProps = (excludeProps: string[] = []) => (state: any): any => Object.fromEntries(
+  Object.entries(state).filter(([key]) => !excludeProps.includes(key)))
 
 export const useSingleViewStore = create<
   SingleViewStore,
@@ -19,10 +24,15 @@ export const useSingleViewStore = create<
 >(
   persist((set) => ({
   lastId: '',
+  lastIndex: -1,
   showDetails: false,
   showNavigation: true,
 
-  setLastId: (lastId: string) => set((state) => ({...state, lastId})),
-  setShowDetails: (show: boolean) => set((state) => ({...state, showDetails: show})),
-  setShowNavigation: (show: boolean) => set((state) => ({...state, showNavigation: show})),
-}), { name: 'gallery-single-view' }))
+  setLastId: (lastId: string) => set((state: SingleViewStore) => ({...state, lastId})),
+  setLastIndex: (lastIndex: number) => set((state: SingleViewStore) => ({...state, lastIndex})),
+  setShowDetails: (show: boolean) => set((state: SingleViewStore) => ({...state, showDetails: show})),
+  setShowNavigation: (show: boolean) => set((state: SingleViewStore) => ({...state, showNavigation: show})),
+}), {
+  name: 'gallery-single-view',
+  partialize: excludeStateProps(['lastId', 'lastIndex']),
+}))